updated according to deeplabv3 choices
